MemCardFile.cpp: Use QImage* for the banner and icon.
This seems to avoid the weird problems encountered with QImage being
initialized incorrectly, resulting in segmentation faults.

MemCard.cpp: Minor fixes.

card.h: Split CARD_ICON_CI into CARD_ICON_CI_UNIQUE and CARD_ICON_CI_SHARED.
It's possible for CI8 icons to share palettes, which currently isn't
supported due to the way GcImage::FromCI8() works.

+++ src/mcrecover/MemCard.hpp
@@ -78,7 +78,7 @@ class MemCard : public QObject
 		 * @param blockIdx Block index.
 		 * @return Bytes read on success; negative on error.
 		 */
-		int readBlock(void *buf, size_t siz, uint16_t blockIdx);
+		int readBlock(void *buf, int siz, uint16_t blockIdx);
 		
 		/**
 		 * Get the memory card encoding.
